Refactor sensitivity scale functions to support continuous float values (1.0-5.0)
- Updated scale_to_pitch_threshold, scale_to_distance_threshold, scale_to_head_roll_threshold, and scale_to_shoulder_tilt_threshold functions to accept float values instead of integers.
- Introduced _interpolate_threshold function for linear interpolation between discrete scale points.
- Modified update_thresholds_from_scales to accept float parameters.
- Adjusted websocket_server to handle float sensitivity scales when setting thresholds.

+++ dotnet-app/Services/JsonContext.cs
@@ -34,10 +34,10 @@ public class SavePostureMessage
 public class SetThresholdsMessage
 {
     public string type { get; set; } = "set_thresholds";
-    public int pitch_scale { get; set; }
-    public int distance_scale { get; set; }
-    public int head_roll_scale { get; set; }
-    public int shoulder_tilt_scale { get; set; }
+    public double pitch_scale { get; set; }
+    public double distance_scale { get; set; }
+    public double head_roll_scale { get; set; }
+    public double shoulder_tilt_scale { get; set; }
 }
 
 public class GetStatisticsMessage
